refactored evaluateHand and turn methods, added tests for untested hand evaluater methods
