+++ src/main/java/ch/uzh/ifi/hase/soprafs24/entity/Game.java
@@ -4,7 +4,7 @@ package ch.uzh.ifi.hase.soprafs24.entity;
 import ch.uzh.ifi.hase.soprafs24.constant.Hand;
 import ch.uzh.ifi.hase.soprafs24.helpers.Card;
 import ch.uzh.ifi.hase.soprafs24.helpers.DeckOfCardsApi;
-import ch.uzh.ifi.hase.soprafs24.helpers.PlayerHand;
+import ch.uzh.ifi.hase.soprafs24.hand.PlayerHand;
 import org.springframework.http.HttpStatus;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
@@ -218,6 +218,5 @@ public class Game implements Serializable {
         setsNextPlayerTurnIndex();
         setsNextPlayerTurnIndex();
         //GameService.startTimer(this.id, this.players.get(this.playerTurnIndex).getToken());
-
     }
 }
