+++ src/main/java/ch/uzh/ifi/hase/soprafs24/hand/CheckHand/CheckHand.java
@@ -1,0 +1,11 @@
+package ch.uzh.ifi.hase.soprafs24.hand.CheckHand;
+
+import ch.uzh.ifi.hase.soprafs24.entity.Player;
+import ch.uzh.ifi.hase.soprafs24.hand.PlayerHand;
+import ch.uzh.ifi.hase.soprafs24.helpers.Card;
+
+import java.util.List;
+
+public interface CheckHand {
+    PlayerHand checkHand(List<Card> cards, Player player);
+}
